refactor(parser): extract shared clearTransEncoding helper
Both WebApParser and StdsysParser carried byte-for-byte identical copies
of clearTransEncoding (35 lines each) to strip leaked HTTP chunked
transfer-encoding markers before UTF-8 decoding. Hoist the routine to a
top-level function in parser_utils.dart and import it at both call sites.

Addresses the "抽取共用工具方法" checklist item in #230.
